requirements.txt
streamlit
PyPDF2
python-docx
langchain
langchain-groq
langchain-community
langchain-huggingface
faiss-cpu
python-dotenv
transformers==4.49.0
huggingface-hub
sentence-transformers
streamlit-chat
pinecone
langchain-pinecone
fastapi
uvicorn
python-docx
langchain-core
pypdf
langgraph